Reorganization, Clean up, Notes
Notated the .htaccess file for better descriptions
Moved code from the application specific INIT file to the API framework
app class. This is code that should not be touched on a regular basis
because it contains core functionality.
Removed the class.app.init.php file as a result of the core class merge
Renamed the class.app.extension.php to class.app.extends.php
Moved some methods from the frameworks app class to the
class.app.extends.php class. These are methods that are
editable/removable and do not affect core functionality.